DeepLab_background_removal_model_development_and_flask_deployment_on_AWS_EC2/requirements.txt
torch
tensorflow
PyOpenGL
PyOpenGL_accelerate
urllib3
opencv-python
pillow
torchvision
torchaudio
numpy 
flask